+++ packages/button-react/example/index.tsx
@@ -1,8 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { PrimaryButton, SecondaryButton, TertiaryButton } from "../src";
+import { initTabListener } from "@fremtind/jkl-core";
 import "@fremtind/jkl-button/button.css";
-import "@fremtind/jkl-core/core.min.css";
+import "@fremtind/jkl-core/core.css";
+
+initTabListener();
 
 function onClick() {
     console.log("Hello!");
@@ -10,14 +13,31 @@ function onClick() {
 
 const Buttons = () => (
     <>
-        <div style={{ margin: "20px" }}>
-            <PrimaryButton onClick={onClick}>Hello, Primary Button</PrimaryButton>
+        <div className="side-by-side">
+            <pre>
+                <code>{`forceCompact={false}`}</code>
+            </pre>
+            <pre>
+                <code>{`forceCompact={true}`}</code>
+            </pre>
         </div>
-        <div style={{ margin: "20px" }}>
-            <SecondaryButton onClick={onClick}>Hello, Secondary Button</SecondaryButton>
+        <div className="side-by-side">
+            <PrimaryButton onClick={onClick}>Send inn</PrimaryButton>
+            <PrimaryButton forceCompact onClick={onClick}>
+                Send inn
+            </PrimaryButton>
         </div>
-        <div style={{ margin: "20px" }}>
-            <TertiaryButton onClick={onClick}>Tertiary</TertiaryButton>
+        <div className="side-by-side">
+            <SecondaryButton onClick={onClick}>Svar senere</SecondaryButton>
+            <SecondaryButton forceCompact onClick={onClick}>
+                Svar senere
+            </SecondaryButton>
+        </div>
+        <div className="side-by-side">
+            <TertiaryButton onClick={onClick}>Avbryt</TertiaryButton>
+            <TertiaryButton forceCompact onClick={onClick}>
+                Avbryt
+            </TertiaryButton>
         </div>
     </>
 );