+++ packages/core/src/index.ts
@@ -1,3 +1,7 @@
-export { initTabListener } from "./utils";
+export { initTabListener, getValuePair } from "./utils";
 
-export type LabelVariant = "secondary" | "small";
+export type LabelVariant = "small" | "medium" | "large";
+export type ValuePair = {
+    value: string;
+    label: string;
+};