+++ packages/datepicker-react/src/DatePicker.tsx
@@ -1,5 +1,5 @@
 import React, { ChangeEvent, useRef, useState } from "react";
-import { SupportLabel } from "@fremtind/jkl-typography-react";
+import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
 import { useClickOutside, useFocusOutside, useKeyListener } from "@fremtind/jkl-react-hooks";
 import { LabelVariant } from "@fremtind/jkl-core";
 // @ts-ignore
@@ -18,11 +18,13 @@ interface Props {
     days?: string[];
     initialDate?: Date;
     onChange?: (date: Date) => void;
+    extended?: boolean;
     initialShow?: boolean;
     className?: string;
     helpLabel?: string;
     errorLabel?: string;
     variant?: LabelVariant;
+    forceCompact?: boolean;
 }
 
 const dayMonthYearRegex = /^(\d\d)\.(\d\d)\.(\d{4})/;
@@ -55,15 +57,26 @@ export function DatePicker({
     days,
     initialDate,
     onChange,
+    extended = false,
     initialShow = false,
     className = "",
     errorLabel,
     helpLabel,
     variant,
+    forceCompact,
 }: Props) {
     const [date, setDate] = useState(initialDate);
     const [datepickerHidden, setDatepickerHidden] = useState(!initialShow);
     const [dateString, setDateString] = useState(initialDate ? formatDate(initialDate) : "");
+    const componentClassName = "jkl-datepicker".concat(
+        extended ? " jkl-datepicker--extended" : "",
+        !datepickerHidden ? " jkl-datepicker--open" : "",
+        className ? ` ${className}` : "",
+    );
+    const inputClassName = "jkl-text-field jkl-datepicker__input".concat(
+        forceCompact ? ` jkl-text-field--compact` : "",
+    );
+
     const openDatepicker = (e: React.FocusEvent<HTMLInputElement>) => {
         // Workaround for loosing focus when opening in chrome:
         // https://github.com/nrkno/core-components/issues/322
@@ -112,10 +125,12 @@ export function DatePicker({
     }
 
     return (
-        <div className={`jkl-datepicker ${className}`} ref={datepickerRef}>
+        <div className={componentClassName} ref={datepickerRef}>
             <div className="jkl-datepicker__outer-wrapper">
-                <label className={`jkl-text-field`}>
-                    <span className={`jkl-label ${variant ? `jkl-label--${variant}` : ""}`}>{label}</span>
+                <label className={inputClassName}>
+                    <Label variant={variant} forceCompact={forceCompact}>
+                        {label}
+                    </Label>
                     <input
                         placeholder={placeholder}
                         type="text"
@@ -136,23 +151,45 @@ export function DatePicker({
                         onDatepickerClickDay={onClickCalendarDay}
                         className="jkl-datepicker__calendar"
                     >
-                        <div className="jkl-datepicker__calendar-navigation">
-                            <label className="jkl-text-field jkl-datepicker__year-selector">
-                                <span className="jkl-label jkl-label--small">{yearLabel}</span>
-                                <input type="year" className={`jkl-text-field__input`} />
-                            </label>
+                        {extended && (
+                            <div className="jkl-datepicker__calendar-navigation">
+                                <label className="jkl-text-field jkl-datepicker__year-selector">
+                                    <Label variant="small">{yearLabel}</Label>
+                                    <input type="year" className={`jkl-text-field__input`} />
+                                </label>
 
-                            <label className="jkl-dropdown jkl-datepicker__month-selector">
-                                <span className="jkl-label jkl-label--small">{monthLabel}</span>
-                                <select className="jkl-dropdown__value"></select>
-                                <span className="jkl-dropdown__chevron" />
-                            </label>
-                        </div>
+                                <label className="jkl-dropdown jkl-datepicker__month-selector">
+                                    <Label variant="small">{monthLabel}</Label>
+                                    <select className="jkl-dropdown__value"></select>
+                                    <span className="jkl-dropdown__chevron" />
+                                </label>
+                            </div>
+                        )}
+                        {!extended && (
+                            <fieldset className="jkl-datepicker__month-navigation">
+                                <button
+                                    title="forrige måned"
+                                    aria-label="forrige måned"
+                                    className="jkl-datepicker__month-button"
+                                    value="- 1 month"
+                                >
+                                    <span className="jkl-datepicker__month-arrow jkl-datepicker__month-arrow--left" />
+                                </button>
+                                <button
+                                    title="neste måned"
+                                    aria-label="neste måned"
+                                    className="jkl-datepicker__month-button jkl-datepicker__month-button--right"
+                                    value="+ 1 month"
+                                >
+                                    <span className="jkl-datepicker__month-arrow jkl-datepicker__month-arrow--right" />
+                                </button>
+                            </fieldset>
+                        )}
                         <table data-testid="jkl-datepicker-calendar" />
                     </CoreDatepicker>
                 </div>
             </div>
-            <SupportLabel errorLabel={errorLabel} helpLabel={helpLabel} />
+            <SupportLabel errorLabel={errorLabel} helpLabel={helpLabel} forceCompact={forceCompact} />
         </div>
     );
 }