+++ packages/field-group-react/src/FieldGroup.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import { SupportLabel } from "@fremtind/jkl-typography-react";
+import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
 import { LabelVariant } from "@fremtind/jkl-core";
 
 interface Props {
@@ -9,14 +9,26 @@ interface Props {
     helpLabel?: string;
     errorLabel?: string;
     variant?: LabelVariant;
+    forceCompact?: boolean;
 }
 
-export const FieldGroup = ({ legend, className, children, helpLabel, errorLabel, variant }: Props) => {
+export const FieldGroup = ({
+    legend,
+    className,
+    children,
+    helpLabel,
+    errorLabel,
+    variant = "medium",
+    forceCompact,
+}: Props) => {
     const componentClassName = "jkl-field-group".concat(className ? ` ${className}` : "");
-    const labelClassName = "jkl-label".concat(variant ? ` jkl-label--${variant}` : "");
     return (
         <fieldset className={componentClassName}>
-            <legend className={labelClassName}>{legend}</legend>
+            <legend>
+                <Label variant={variant} forceCompact={forceCompact}>
+                    {legend}
+                </Label>
+            </legend>
             {children}
             {(helpLabel || errorLabel) && <SupportLabel helpLabel={helpLabel} errorLabel={errorLabel} />}
         </fieldset>