+++ packages/radio-button-react/src/RadioButtonOption.tsx
@@ -1,18 +1,30 @@
-import React from "react";
+import React, { ChangeEventHandler } from "react";
 
 interface Props {
     value: string;
+    label: string;
     name: string;
-    onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+    onChange: ChangeEventHandler<HTMLInputElement>;
     inline?: boolean;
     checked?: boolean;
     invalid?: boolean;
+    forceCompact?: boolean;
 }
 
-export const RadioButton = ({ value, onChange, name = "", inline = false, checked, invalid = false }: Props) => (
+export const RadioButtonOption = ({
+    value,
+    label,
+    onChange,
+    name = "",
+    inline = false,
+    checked,
+    invalid = false,
+    forceCompact,
+}: Props) => (
     <label
         data-testid="jkl-radio-button__label-tag"
         className={"jkl-radio-button".concat(
+            forceCompact ? " jkl-radio-button--compact" : "",
             inline ? " jkl-radio-button--inline" : "",
             invalid ? " jkl-radio-button--error" : "",
         )}
@@ -27,6 +39,6 @@ export const RadioButton = ({ value, onChange, name = "", inline = false, checke
             aria-invalid={invalid}
         />
         <span className="jkl-radio-button__dot" />
-        <span className="jkl-radio-button__label">{value}</span>
+        <span className="jkl-radio-button__label">{label}</span>
     </label>
 );