+++ packages/text-input-react/src/TextArea.tsx
@@ -1,6 +1,6 @@
-import React, { ChangeEvent, useState, FocusEvent } from "react";
+import React, { ChangeEvent, FocusEvent } from "react";
 import { LabelVariant } from "@fremtind/jkl-core";
-import { SupportLabel } from "@fremtind/jkl-typography-react";
+import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
 
 interface Props {
     label: string;
@@ -15,6 +15,7 @@ interface Props {
     rows?: number;
     placeholder?: string;
     variant?: LabelVariant;
+    forceCompact?: boolean;
 }
 
 export const TextArea = ({
@@ -26,37 +27,27 @@ export const TextArea = ({
     errorLabel,
     rows = 7,
     placeholder = " ",
+    forceCompact,
     ...restProps
 }: Props) => {
-    const [isFocused, setIsFocused] = useState(false);
-
-    const height = isFocused || restProps.value ? `${rows * 2 + 0.5}rem` : undefined;
-
-    const componentClassName = "jkl-text-field jkl-text-area".concat(className ? ` ${className}` : "");
-    const labelClassName = "jkl-label".concat(variant ? ` jkl-label--${variant}` : "");
-    function onBlur(event: FocusEvent<HTMLTextAreaElement>) {
-        setIsFocused(false);
-        restProps.onBlur && restProps.onBlur(event);
-    }
-
-    function onFocus() {
-        return setIsFocused(true);
-    }
+    const componentClassName = "jkl-text-field jkl-text-area".concat(
+        forceCompact ? ` jkl-text-field--compact` : "",
+        className ? ` ${className}` : "",
+    );
 
     return (
         <label data-testid="jkl-text-field" className={componentClassName}>
-            <span className={labelClassName}>{label}</span>
+            <Label variant={variant} forceCompact={forceCompact}>
+                {label}
+            </Label>
             <textarea
-                onFocus={onFocus}
-                onBlur={onBlur}
-                style={{ height }}
                 aria-invalid={!!errorLabel}
-                className="jkl-text-field__input"
+                className={`jkl-text-field__input jkl-text-field__input--${rows}-rows`}
                 id={id}
                 placeholder={placeholder}
                 {...restProps}
             />
-            <SupportLabel helpLabel={helpLabel} errorLabel={errorLabel} />
+            <SupportLabel helpLabel={helpLabel} errorLabel={errorLabel} forceCompact={forceCompact} />
         </label>
     );
 };