+++ portal/src/pages/components/card.tsx
@@ -5,11 +5,8 @@ import { ComponentLayout, Changelog } from "../../components";
 import { renderer } from "../../presentation/markdownRenderer";
 import { CardExample } from "../../examples";
 
-// @ts-ignore
 import readmeContents from "!raw-loader!@fremtind/jkl-card/README.md";
-// @ts-ignore
 import changelog from "!raw-loader!@fremtind/jkl-card/CHANGELOG.md";
-// @ts-ignore
 import changelogReact from "!raw-loader!@fremtind/jkl-card-react/CHANGELOG.md";
 
 export default function CardPage() {