+++ portal/src/pages/developer/test.tsx
@@ -3,7 +3,6 @@ import ReactMarkdown from "react-markdown";
 import { Layout } from "../../components";
 import { renderer } from "../../presentation/markdownRenderer";
 
-// @ts-ignore
 import TestGuideMarkdown from "!raw-loader!../../../../guides/TestGuide.md";
 
 const GettingStarted = () => {