+++ packages/field-group-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-field-group-react@0.2.3...@fremtind/jkl-field-group-react@0.3.0) (2019-10-09)
+
+
+### Features
+
+* **field-group:** add Field component for single input fields ([4457804](https://github.com/fremtind/jokul/commit/4457804))
+* **field-group:** implement new design of field group labels ([6df2939](https://github.com/fremtind/jokul/commit/6df2939))
+* **field-group:** remove Field component and update style of FieldGroup ([8a451ac](https://github.com/fremtind/jokul/commit/8a451ac))
+* **field-group:** update field group styling according to sketches ([69d94d8](https://github.com/fremtind/jokul/commit/69d94d8))
+* change name of label variants ([554749c](https://github.com/fremtind/jokul/commit/554749c))
+* extract label styles to core ([9fad287](https://github.com/fremtind/jokul/commit/9fad287))
+
+
+
+
+
 ## [0.2.3](https://github.com/fremtind/jokul/compare/@fremtind/jkl-field-group-react@0.2.2...@fremtind/jkl-field-group-react@0.2.3) (2019-09-26)
 
 **Note:** Version bump only for package @fremtind/jkl-field-group-react