+++ packages/message-box/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.7](https://github.com/fremtind/jokul/compare/@fremtind/jkl-message-box@0.7.6...@fremtind/jkl-message-box@0.7.7) (2019-10-09)
+
+**Note:** Version bump only for package @fremtind/jkl-message-box
+
+
+
+
+
 ## [0.7.6](https://github.com/fremtind/jokul/compare/@fremtind/jkl-message-box@0.7.5...@fremtind/jkl-message-box@0.7.6) (2019-09-26)
 
 **Note:** Version bump only for package @fremtind/jkl-message-box